Refactor PolicyEngine to Core Package (#12325)
