+++ packages/cli/src/services/BuiltinCommandLoader.ts
@@ -28,6 +28,7 @@ import { memoryCommand } from '../ui/commands/memoryCommand.js';
 import { modelCommand } from '../ui/commands/modelCommand.js';
 import { permissionsCommand } from '../ui/commands/permissionsCommand.js';
 import { privacyCommand } from '../ui/commands/privacyCommand.js';
+import { policiesCommand } from '../ui/commands/policiesCommand.js';
 import { profileCommand } from '../ui/commands/profileCommand.js';
 import { quitCommand } from '../ui/commands/quitCommand.js';
 import { restoreCommand } from '../ui/commands/restoreCommand.js';
@@ -75,6 +76,9 @@ export class BuiltinCommandLoader implements ICommandLoader {
       ...(this.config?.getUseModelRouter() ? [modelCommand] : []),
       ...(this.config?.getFolderTrust() ? [permissionsCommand] : []),
       privacyCommand,
+      ...(this.config?.getEnableMessageBusIntegration()
+        ? [policiesCommand]
+        : []),
       ...(isDevelopment ? [profileCommand] : []),
       quitCommand,
       restoreCommand(this.config),
