+++ packages/core/src/config/config.test.ts
@@ -13,11 +13,11 @@ import type {
 } from './config.js';
 import {
   Config,
-  ApprovalMode,
   DEFAULT_FILE_FILTERING_OPTIONS,
   HookType,
   HookEventName,
 } from './config.js';
+import { ApprovalMode } from '../policy/types.js';
 import * as path from 'node:path';
 import { setGeminiMdFilename as mockSetGeminiMdFilename } from '../tools/memoryTool.js';
 import {
