+++ packages/core/src/policy/index.ts
@@ -6,3 +6,5 @@
 
 export * from './policy-engine.js';
 export * from './types.js';
+export * from './toml-loader.js';
+export * from './config.js';
