+++ packages/core/src/policy/toml-loader.test.ts
@@ -5,7 +5,7 @@
  */
 
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
-import { ApprovalMode, PolicyDecision } from '@google/gemini-cli-core';
+import { ApprovalMode, PolicyDecision } from './types.js';
 import type { Dirent } from 'node:fs';
 import nodePath from 'node:path';
 
@@ -66,9 +66,7 @@ priority = 100
         readdir: mockReaddir,
       }));
 
-      const { loadPoliciesFromToml: load } = await import(
-        './policy-toml-loader.js'
-      );
+      const { loadPoliciesFromToml: load } = await import('./toml-loader.js');
 
       const getPolicyTier = (_dir: string) => 1;
       const result = await load(
@@ -133,9 +131,7 @@ priority = 100
         readdir: mockReaddir,
       }));
 
-      const { loadPoliciesFromToml: load } = await import(
-        './policy-toml-loader.js'
-      );
+      const { loadPoliciesFromToml: load } = await import('./toml-loader.js');
 
       const getPolicyTier = (_dir: string) => 2;
       const result = await load(
@@ -203,9 +199,7 @@ priority = 100
         readdir: mockReaddir,
       }));
 
-      const { loadPoliciesFromToml: load } = await import(
-        './policy-toml-loader.js'
-      );
+      const { loadPoliciesFromToml: load } = await import('./toml-loader.js');
 
       const getPolicyTier = (_dir: string) => 2;
       const result = await load(
@@ -273,9 +267,7 @@ priority = 100
         readdir: mockReaddir,
       }));
 
-      const { loadPoliciesFromToml: load } = await import(
-        './policy-toml-loader.js'
-      );
+      const { loadPoliciesFromToml: load } = await import('./toml-loader.js');
 
       const getPolicyTier = (_dir: string) => 1;
       const result = await load(
@@ -340,9 +332,7 @@ priority = 100
         readdir: mockReaddir,
       }));
 
-      const { loadPoliciesFromToml: load } = await import(
-        './policy-toml-loader.js'
-      );
+      const { loadPoliciesFromToml: load } = await import('./toml-loader.js');
 
       const getPolicyTier = (_dir: string) => 2;
       const result = await load(
@@ -410,9 +400,7 @@ modes = ["yolo"]
         readdir: mockReaddir,
       }));
 
-      const { loadPoliciesFromToml: load } = await import(
-        './policy-toml-loader.js'
-      );
+      const { loadPoliciesFromToml: load } = await import('./toml-loader.js');
 
       const getPolicyTier = (_dir: string) => 1;
       const result = await load(
@@ -473,9 +461,7 @@ priority = 100
         readdir: mockReaddir,
       }));
 
-      const { loadPoliciesFromToml: load } = await import(
-        './policy-toml-loader.js'
-      );
+      const { loadPoliciesFromToml: load } = await import('./toml-loader.js');
 
       const getPolicyTier = (_dir: string) => 1;
       const result = await load(
@@ -535,9 +521,7 @@ priority = 100
         readdir: mockReaddir,
       }));
 
-      const { loadPoliciesFromToml: load } = await import(
-        './policy-toml-loader.js'
-      );
+      const { loadPoliciesFromToml: load } = await import('./toml-loader.js');
 
       const getPolicyTier = (_dir: string) => 1;
       const result = await load(
@@ -599,9 +583,7 @@ priority = 100
         readdir: mockReaddir,
       }));
 
-      const { loadPoliciesFromToml: load } = await import(
-        './policy-toml-loader.js'
-      );
+      const { loadPoliciesFromToml: load } = await import('./toml-loader.js');
 
       const getPolicyTier = (_dir: string) => 1;
       const result = await load(
@@ -663,9 +645,7 @@ priority = 100
         readdir: mockReaddir,
       }));
 
-      const { loadPoliciesFromToml: load } = await import(
-        './policy-toml-loader.js'
-      );
+      const { loadPoliciesFromToml: load } = await import('./toml-loader.js');
 
       const getPolicyTier = (_dir: string) => 1;
       const result = await load(
@@ -726,9 +706,7 @@ priority = 100
         readdir: mockReaddir,
       }));
 
-      const { loadPoliciesFromToml: load } = await import(
-        './policy-toml-loader.js'
-      );
+      const { loadPoliciesFromToml: load } = await import('./toml-loader.js');
 
       const getPolicyTier = (_dir: string) => 1;
       const result = await load(
@@ -790,9 +768,7 @@ priority = 100
         readdir: mockReaddir,
       }));
 
-      const { loadPoliciesFromToml: load } = await import(
-        './policy-toml-loader.js'
-      );
+      const { loadPoliciesFromToml: load } = await import('./toml-loader.js');
 
       const getPolicyTier = (_dir: string) => 1;
       const result = await load(
@@ -802,181 +778,14 @@ priority = 100
       );
 
       expect(result.rules).toHaveLength(1);
-      // Should match literal asterisk, not wildcard
+      // The regex should have escaped the * and .
+      expect(
+        result.rules[0].argsPattern?.test('{"command":"git log file.txt"}'),
+      ).toBe(false);
       expect(
         result.rules[0].argsPattern?.test('{"command":"git log *.txt"}'),
       ).toBe(true);
-      expect(
-        result.rules[0].argsPattern?.test('{"command":"git log a.txt"}'),
-      ).toBe(false);
       expect(result.errors).toHaveLength(0);
-    });
-
-    it('should handle non-existent directory gracefully', async () => {
-      const actualFs =
-        await vi.importActual<typeof import('node:fs/promises')>(
-          'node:fs/promises',
-        );
-
-      const mockReaddir = vi.fn(async (_path: string): Promise<Dirent[]> => {
-        const error: NodeJS.ErrnoException = new Error('ENOENT');
-        error.code = 'ENOENT';
-        throw error;
-      });
-
-      vi.doMock('node:fs/promises', () => ({
-        ...actualFs,
-        default: { ...actualFs, readdir: mockReaddir },
-        readdir: mockReaddir,
-      }));
-
-      const { loadPoliciesFromToml: load } = await import(
-        './policy-toml-loader.js'
-      );
-
-      const getPolicyTier = (_dir: string) => 1;
-      const result = await load(
-        ApprovalMode.DEFAULT,
-        ['/non-existent'],
-        getPolicyTier,
-      );
-
-      // Should not error for missing directories
-      expect(result.rules).toHaveLength(0);
-      expect(result.errors).toHaveLength(0);
-    });
-
-    it('should reject priority >= 1000 with helpful error message', async () => {
-      const actualFs =
-        await vi.importActual<typeof import('node:fs/promises')>(
-          'node:fs/promises',
-        );
-
-      const mockReaddir = vi.fn(
-        async (
-          path: string,
-          _options?: { withFileTypes: boolean },
-        ): Promise<Dirent[]> => {
-          if (nodePath.normalize(path) === nodePath.normalize('/policies')) {
-            return [
-              {
-                name: 'invalid.toml',
-                isFile: () => true,
-                isDirectory: () => false,
-              } as Dirent,
-            ];
-          }
-          return [];
-        },
-      );
-
-      const mockReadFile = vi.fn(async (path: string): Promise<string> => {
-        if (
-          nodePath.normalize(path) ===
-          nodePath.normalize(nodePath.join('/policies', 'invalid.toml'))
-        ) {
-          return `
-[[rule]]
-toolName = "glob"
-decision = "allow"
-priority = 1000
-`;
-        }
-        throw new Error('File not found');
-      });
-
-      vi.doMock('node:fs/promises', () => ({
-        ...actualFs,
-        default: { ...actualFs, readFile: mockReadFile, readdir: mockReaddir },
-        readFile: mockReadFile,
-        readdir: mockReaddir,
-      }));
-
-      const { loadPoliciesFromToml: load } = await import(
-        './policy-toml-loader.js'
-      );
-
-      const getPolicyTier = (_dir: string) => 1;
-      const result = await load(
-        ApprovalMode.DEFAULT,
-        ['/policies'],
-        getPolicyTier,
-      );
-
-      expect(result.rules).toHaveLength(0);
-      expect(result.errors).toHaveLength(1);
-      expect(result.errors[0].errorType).toBe('schema_validation');
-      expect(result.errors[0].details).toContain('priority');
-      expect(result.errors[0].details).toContain('tier overflow');
-      expect(result.errors[0].details).toContain(
-        'Priorities >= 1000 would jump to the next tier',
-      );
-      expect(result.errors[0].details).toContain('<= 999');
-    });
-
-    it('should reject negative priority with helpful error message', async () => {
-      const actualFs =
-        await vi.importActual<typeof import('node:fs/promises')>(
-          'node:fs/promises',
-        );
-
-      const mockReaddir = vi.fn(
-        async (
-          path: string,
-          _options?: { withFileTypes: boolean },
-        ): Promise<Dirent[]> => {
-          if (nodePath.normalize(path) === nodePath.normalize('/policies')) {
-            return [
-              {
-                name: 'invalid.toml',
-                isFile: () => true,
-                isDirectory: () => false,
-              } as Dirent,
-            ];
-          }
-          return [];
-        },
-      );
-
-      const mockReadFile = vi.fn(async (path: string): Promise<string> => {
-        if (
-          nodePath.normalize(path) ===
-          nodePath.normalize(nodePath.join('/policies', 'invalid.toml'))
-        ) {
-          return `
-[[rule]]
-toolName = "glob"
-decision = "allow"
-priority = -1
-`;
-        }
-        throw new Error('File not found');
-      });
-
-      vi.doMock('node:fs/promises', () => ({
-        ...actualFs,
-        default: { ...actualFs, readFile: mockReadFile, readdir: mockReaddir },
-        readFile: mockReadFile,
-        readdir: mockReaddir,
-      }));
-
-      const { loadPoliciesFromToml: load } = await import(
-        './policy-toml-loader.js'
-      );
-
-      const getPolicyTier = (_dir: string) => 1;
-      const result = await load(
-        ApprovalMode.DEFAULT,
-        ['/policies'],
-        getPolicyTier,
-      );
-
-      expect(result.rules).toHaveLength(0);
-      expect(result.errors).toHaveLength(1);
-      expect(result.errors[0].errorType).toBe('schema_validation');
-      expect(result.errors[0].details).toContain('priority');
-      expect(result.errors[0].details).toContain('>= 0');
-      expect(result.errors[0].details).toContain('must be >= 0');
     });
   });
 });
