+++ packages/core/src/tools/smart-edit.test.ts
@@ -53,7 +53,8 @@ import { ToolErrorType } from './tool-error.js';
 import path from 'node:path';
 import fs from 'node:fs';
 import os from 'node:os';
-import { ApprovalMode, type Config } from '../config/config.js';
+import { ApprovalMode } from '../policy/types.js';
+import { type Config } from '../config/config.js';
 import { type Content, type Part, type SchemaUnion } from '@google/genai';
 import { createMockWorkspaceContext } from '../test-utils/mockWorkspaceContext.js';
 import { StandardFileSystemService } from '../services/fileSystemService.js';
